Make daemon errors more useful.  This makes the errinfo expected by the caller adhere to the convention the async jobret function expects, so more verbose error information is made available. This is added to all the job completion routines the dmon runs and hands back to telepath. The getDmonCore function from test_cmds_cortex is moved into the SynTest class outright for reuse.
